TASK 72: display thumbnail and play embedded video in list
Grabs thumbnails generated by the FunctionVideoThumbnail Azure function from the videocontainer and correlates the thumbnails with matching videos. Each thumbnail is clickable and on-click the screen is updated to show the desired video embedded in the list.

Refreshes the screen after video upload and deletion. Handles the case in which a video thumbnail is not present for a video. On deletion, cleans up in memory structures and Azure blob storage.