ISSUE 83: OPSWAT file scanning
Prior to uploading video files to Azure storage container, the video is now first scanned by the OPSWAT threat prevention platform. This is done via two HTTP requests to its /file endpoint:

1) POST - sends file to the OPSWAT service and returns a dataId to track the status of the scan.

2) GET - probes for the result of the scan using the dataId provided in response to POST. Also returns a progress percentage.

Added status to the UI to give user some feedback whilst the video is being uploaded to scanning service, the scan is in progress and the video is being uploaded to Azure. Fixed the spinner.

Authentication is achieved using an API key, this key is currently being passed in via appsettings.json like other services.