TASK 184: update name available check to call new API endpoint.
Moved name available logic into the API to prevent returning more information than required with a GetVideos(brokerId: 0) call. Submitting a request with a brokerId of 0 also fails the VerifyBroker() checks in the API. Moving the name available logic into the API means we can validate the logged in broker and only return a true/false value as to whether or not the filename is available, which is all we should return.