ISSUE 252: Chat Templates
Allow users to select a template from a dropdown in the chat tab of the webpage. Templates include three dynamic fields - INSERT_CLIENT_NAME, INSERT_BROKER_NAME and INSERT_DATE. The first two are replaced automatically when a client is selected whilst INSERT_DATE must be filled in by user before sending a template.

Users can edit templates via the edit templates button in the chat tab. On the chattemplateedit page, users can click on each entry in the table and modify the contents. User can cancel their edit which reset the display, or save their changes which will update the broker defined messages in the database.